Delete endpoint now validates if todo you want to delete is yours

+++ TodoApp/main.py
@@ -51,12 +51,15 @@ async def read_todo(todo_id: int, user: dict = Depends(get_current_user), db: Se
 
 
 @app.post("/")
-async def create_todo(todo: Todo, db: Session = Depends(get_db)):
+async def create_todo(todo: Todo, user: dict = Depends(get_current_user), db: Session = Depends(get_db)):
+    if not user:
+        raise get_user_exception()
     todo_model = models.Todos()
     todo_model.title = todo.title
     todo_model.description = todo.description
     todo_model.priority = todo.priority
     todo_model.complete = todo.complete
+    todo_model.owner_id = user.get("id")
 
     db.add(todo_model)
     db.commit()
@@ -65,8 +68,11 @@ async def create_todo(todo: Todo, db: Session = Depends(get_db)):
 
 
 @app.put("/{todo_id}")
-async def update_todo(todo_id: int, todo: Todo, db: Session = Depends(get_db)):
-    todo_model = db.query(models.Todos).filter(models.Todos.id == todo_id).first()
+async def update_todo(todo_id: int, todo: Todo, user: dict = Depends(get_current_user), db: Session = Depends(get_db)):
+    if not user:
+        raise get_user_exception()
+    todo_model = db.query(models.Todos).filter(models.Todos.id == todo_id).filter(
+        models.Todos.owner_id == user.get("id")).first()
     if not todo_model:
         raise http_exception()
     todo_model.title = todo.title
@@ -80,9 +86,12 @@ async def update_todo(todo_id: int, todo: Todo, db: Session = Depends(get_db)):
     return successful_response(200)
 
 
-@app.delete("/{todo_uid}")
-async def delete_todo(todo_id: int, db: Session = Depends(get_db)):
-    todo_model = db.query(models.Todos).filter(models.Todos.id == todo_id).first()
+@app.delete("/{todo_id}")
+async def delete_todo(todo_id: int, user: dict = Depends(get_current_user), db: Session = Depends(get_db)):
+    if not user:
+        raise get_user_exception()
+    todo_model = db.query(models.Todos).filter(models.Todos.id == todo_id).filter(
+        models.Todos.owner_id == user.get("id")).first()
     if not todo_model:
         raise http_exception()
     db.query(models.Todos).filter(models.Todos.id == todo_id).delete()
